Move translations to a shared lib
